+++ tests/capi/doc/chfl_trajectory/topology_file.c
@@ -1,12 +1,15 @@
 #include <chemfiles.h>
+#include <stdlib.h>
 
 int main() {
     // [example] [no-run]
     CHFL_TRAJECTORY* trajectory = chfl_trajectory_open("water.nc", 'r');
 
-    chfl_trajectory_set_topology_file(trajectory, "water.pdb");
-
+    chfl_trajectory_topology_file(trajectory, "water.pdb", NULL);
     /* Reading the trajectory will use topology from water.pdb */
+
+    chfl_trajectory_topology_file(trajectory, "water.topo", "PDB");
+    /* Reading the trajectory will use topology from water.topo using the PDB format. */
 
     chfl_trajectory_close(trajectory);
     // [example]